optimise folder structures
