+++ app/(pages)/how-it-works/page.tsx
@@ -1,9 +1,9 @@
-import HowItWorks from '@/app/components/sections/HowItWorks'
+import HowItWorks from "@/components/sections/HowItWorks";
 
 export default function HowItWorksPage() {
   return (
     <main>
       <HowItWorks />
     </main>
-  )
-} 
+  );
+}
