+++ app/page.tsx
@@ -1,7 +1,9 @@
-import Hero from './components/sections/Hero'
-import WinPrizes from './components/sections/WinPrizes'
-import HowItWorks from './components/sections/HowItWorks'
-import AboutSection from './components/sections/AboutSection'
+import Hero from "@/components/sections/Hero";
+import WinPrizes from "@/components/sections/WinPrizes";
+import HowItWorks from "@/components/sections/HowItWorks";
+import AboutSection from "@/components/sections/AboutSection";
+import FaqSection from "@/components/sections/FaqSection";
+import AnnouncementBanner from "@/components/sections/AnnouncementBanner";
 
 export default function Home() {
   return (
@@ -11,5 +13,5 @@ export default function Home() {
       <HowItWorks />
       <AboutSection />
     </main>
-  )
+  );
 }
