+++ components/sections/AnnouncementBanner.tsx
