+++ components/sections/Hero.tsx
