+++ components/sections/HowItWorks.tsx
